Refactored model training script
