+++ requirements.txt
@@ -1,4 +1,3 @@
-python==3.12
 requests==2.32.3
 pandas==2.2.3
 scikit-learn==1.5.2
@@ -8,3 +7,5 @@ statsmodels==0.14.3
 xgboost==2.1.1
 streamlit==1.38.0
 python-dotenv==1.0.1
+joblib~=1.4.2
+numpy~=2.1.1
