+++ src/config.py
@@ -1,0 +1,6 @@
+import os
+
+BASE_DIR = os.path.dirname(os.path.abspath(__file__))
+RES_DIR = os.path.join(BASE_DIR, '../res')
+DATA_DIR = os.path.join(RES_DIR, 'data')
+MODELS_DIR = os.path.join(RES_DIR, 'models')
